pyproject.toml
[project]
name = "cpan123"
version = "0.1.1"
description = "123云盘的python接口文档(非官方)"
readme = "README.md"
authors = [
    { name = "zoushucai", email = "zoushucai@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "json5>=0.10.0",
    "jsonpath>=0.6.0",
    "py3-wget>=1.0.0",
    "pydantic>=2.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.10.0",
    "tenacity>=8.0.0",
]

[project.scripts]
cpan123 = "cpan123:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"


[tool.ruff.lint]
extend-select = ["I", "B"]
[dependency-groups]
dev = [
    "black>=25.1.0",
    "jsonschema>=4.23.0",
    "mkdocs>=1.6.1",
    "mkdocs-autorefs>=1.4.1",
    "mkdocs-get-deps>=0.2.0",
    "mkdocs-material>=9.6.14",
    "mkdocs-material-extensions>=1.3.1",
    "mkdocstrings>=0.29.1",
    "mkdocstrings-python>=1.16.10",
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]

[project.urls]
Homepage = "https://github.com/zoushucai/cpan123"
Documentation = "https://cpan123.readthedocs.org"
Repository = "https://github.com/zoushucai/cpan123.git"
Issues = "https://github.com/zoushucai/cpan123/issues"
Changelog = "https://github.com/zoushucai/cpan123/blob/main/CHANGELOG.md"
